gce-cluster.sh with no installer and nonpreemptible fixes
gui-installer testing requires the ability to create a GCE without
automatically installing anything on the created cluster.

Changes to gce-cluster.sh do three things:

1. Putting the --no-installer option in place of the installer file
name causes the script to run and create the cluster and use a empty
installer (called noop-installer) to install nothing on it.

2. Testing of the GUI installer without the --preemptible flag
caused errors that sudo needed a tty in the final swap configuration
steps.  The -tt option is added the gcloud compute ssh commands to
provide a pseudo-tty.

3. Running without the --preemptible flag also changes the contents
of gce-output.txt during the "gcloud compute instances create" command.
If --preemptible is specified, a "1" appears in the preemptible output
column.  Without it, nothing is shown and the commands to retrieve
the internal and external IP addresses are off by one column.

Change-Id: I03487df2df3f78aa98958edd2711727937ed72cc