Check in some commonly used functions in Jenkins
There are some functions we use in many of our Jenkins job (e.g.
GCEXCEFuncTest, GCESystemTest, PoseidonTest, etc). We want to start
checking in our Jenkins job into git.

In this first commit, we check in the functions used to control stuff in
GCE. These functions might make it easier for the Jenkins job to be
cloud agnostic and might make it slightly easier to migrate our QA from
GCE into AWS.

Change-Id: I76927f008f012e79db938ad1b10a5d48580f86d3