GuardRails memory allocator/checker.
This commit provides a replacement for the libc heap allocator.  A
logarithmic allocator is provided in which all allocations are tracked,
and a guard page without read/write permissions is appended to the end
of each.  Allocations are made starting from the guard page and
descending in address until satisfied.  This places the allocation
adjacent to the guard page.  Buffer overruns will thus cause a segfault
instead of heap corruption.  Allocations and guard pages are cached and
reused.

Additionally, histograms of allocation statistics are maintained and
leaks are tracked and reported on exit.

Operators and other tests run successfully in this at near native speed.
The allocator is extremely space inefficient with small allocations.

Usage is with LD_PRELOAD.  Also it's probably necessary to increase the
maps limit: echo 10000000 | sudo tee /proc/sys/vm/max_map_count

Also set Constants.NoChildLDPreload=true in XCE config.

This is a first cut; there's room for lots of additional features...

Change-Id: If5fb23cf1b03e1c1d5e255172c3c07be0c1c3c8a