Xc-10108 command-line xcalarctl stop/start does not pick up local Azure storage account - but XD does
The problem here is that /etc/default/xcalar is not readable by xcalar user and
hence xcalarctl run as xcalar user is not able to source the default file.
Because of this the default storage account parameters are not set as env
variables for the usrnode launched by xcalarctl.

This works from XD because:
- expserver has the env variables set and hence are carried over to
xcalarctl and then to usrnode.
- expserver has the env variables set because it was spawned by
supervisord which has these env variables set
- and supervisord has these env variables set as the xcalar service
is started as root which is able to source the file /etc/default/xcalar
and then kick off xcalarctl as xcalar user.

Change-Id: I8fef7058d28907e25bfa4cf09ea5dd3bf1eef3dc