Add -m option to lock/limit memory
In the case of runaway allocations, gr can start swapping.  -m <pct>
takes a maximum percentage of physical memory to use and will cause an
allocator dump followed by abort if that's exceeded.  The result of the
dump should show the problem allocator.

-m also mlocks the memory in question to guarantee against swapping.

Should be run with Constants.EnforceVALimit=false.

Change-Id: If6a32050e4dee257ccbab0aefe4e1c7337260b76