[ovirttool] SSH key to use for SSH to new VMs
When installing Xcalar on a new VM, tool is SSH'ing to the VMs.
Currently, the users public key must be in .ssh/authorized_keys
of the VM, which means it should be in .ssh/authorized_keys of
the template the VM is based from.  This is problematic for any
hosts that call the script, that arise after the template was
created and would require adding those in individually.
This change introduces a public key that the script will pass
when making SSH calls.  The public key will be added to the
existing templates.

Change-Id: I0aa248f3eccefeb378596c0d53296bc7dc88c6db