[ovirttool] fix bug in cmd args passed to python
If user specified --user to ovrittool.sh, the pyton script is
going to get --user --user=<username passed>

This fixes so that --user=<username passed> only will get passed

Also, only prompt for password, if OVIRT_PASSWORD env var is
not set.  This will allow test script to set a password for
duration of process so that shell script won't prompt for it
each test case.

Change-Id: I8f7978d89844027eaa4ec8c1f367effc2f31e97e