Xc-0000 poseidon functest to continue in the wake of ssh errors
If cluster is up and running when a test failed, that usually meaans that
there is some environmental issue terminating the connection between jenkins
slave and the cluster node. In such cases, restart the cluster so that any old
test invocations are killed and then move on to the next test run.

Change-Id: Ide0c75d3ab584d09a5c8833c544fbb87a57841bb