Xc-11588 Jenkins testsuite should use the latest expServer
XDTestsuite uses installer and then just replaces /var/www with the
gui bits that are being tested. The problem with this is that expServer
and jupyter wil be using installer bits and not from the gui changes that
are being tested. This fix will make /opt/xcalar/xcalar-gui to have
the new gui bits and then starts up the cluster. node modules are removed
and rebuilt on the test system to take care of couple things: 1) node version
mismatch between the test system and the gui build system. 2) when change
includes new npm module, then it will be included for the test run

Change-Id: I9da433ca8923d3b79699d499d102079f4a7df82a