XPE; set up data target for $HOME mnt in docker
Context behind this review:
In "Import Data Source", the default selection for the "Data Target"
dropdown menu is "Default Shared Root".  If you click the 'Browse'
button when this default is selected, you will be browsing the root
of the file system where Xcalar is installed.
Because Xcalar is inside a docker container in XPE, then if an XPE
user goes to "Import Data Source", and clicks 'Browse' button on this
default, what they are going to see is the contents of the docker container
itself - not what is local on their file system.  This would be confusing.

This is going to be resolved by
   1. mounting the user's $HOME in to the docker container (Readonly)
   2. creating a data target to that mount point
   3. setting that data target as the default one.

The mount (1.) is done being done as part of this review:
https://gerrit.int.xcalar.com/#/c/9292/

This change takes care of (2.): during the install, once xdpce
conatiner has been created and Xcalar is running (but before XD
is ever launched to the user), a data target is created to /hostmnt
(the mount point their $HOME is mounted to in the container)

An accompanying change in xcalar-gui, will take care of (3.):
That change will be so XD will
check if XPE is being run, and if so, it will display this data target
as the default of the "Data Target" dropdown menu on the
"Import Data Source" page,
(This way when user goes to browse for datasets, if they use the
default data target and click browse, they will see their $HOME dir)

Change-Id: I64f30517179804f0d2e1fa71e9ad9b9c6f64bad7