Dump heap state from core, -a (abort oom) opt
This change add a new GDB macro, gr-dump-in-flight, which can be used to
dump in-flight allocations.  This is mostly useful for core analysis
where a clean shutdown or SIGUSR2 is unavailable to dump the heap state.
Also adds a wrapper script (grcoreheap.sh) to conveniently dump the
state from a core file.

Also adds "-a" option to abort on out of memory.

Change-Id: I0d116d524602657ceaa2dc09e1dd182ed9e4eeab