script to setup gui dir for installer
the xpe .app is currently being generated by the makeapp.sh
script.
One component of creating the .app, is creating the directory
for the installer, and previously this was being done by
makeapp.sh
The new script in this review (createGui.sh) takes care of just
setting up the installer directory, and now makeapp.sh will call
it.
This is being introduced because the installer directory is
standalone, and there is benefit to just building that directory
itself (such as when it is needed to just test the installer out)

In addition, XPE main executable is changed to accomidate a
filename change

Also, XCEINSTALLER_package.json (which is the package.json that
nwjs uses to launch the installer) is changed to accomidate
some requests by Jerene having to do with how the installer
should be displayed by nwjs (such as, it should not be
resizable by the user)

Change-Id: I7a9ced9f3d9f064adc6be083c175711be4d3e6c5