xpe: use xpeInstaller.css from xcalar-gui
before i was copying in a version of the css saved on netstore,
because there was no less file to generate the css so I could
not build it in the xcalar-gui.  However, now there is a proper
less file, so running 'grunt dev' will generate the most current
and needed css file.

Also, remove style.css - it is no longer needed

Change-Id: I1315370dcd4e85bdaf01120443799e9bf37879c2