xpe: Restart app after install to retain app icon
When the app launches, the entrypoint script (docker/xpe/scripts/XPE)
is run.
At this time, the mac dock will display the app icon specified by
the app's Info.plist file.
The entrypoint script runs nwjs as a separate process (either for
installer Gui or XD, depending on the scenario), however, if
at any time the nwjs window is closed, even if the entrypoint script
is still alive and running, mac OS will remove the app icon from the dock.

Previously, for the install scenario, the entrypoint script would
launch nwjs for the installer, and upon successful insall completion,
(at which point the nwjs gui would have closed itself), would
immediately launch another nwjs process for XD, so long as the
'launchfile' is detected (a file placed in the app dir by the
installer server if the user clicks the 'launch' button at the
end of GUI install)
Howver, because of the above, once that initial install nwjs window
closes, the app icon would have disappeared from the mac dock,
and so when the second nwjs process is run, instead the icon for nwjs
would appear.

This review changes the flow of the entrypoint script to avoid this.
This is the new flow for install scenarios

1. nwjs installer gui launched
2. upon successful install comletion, entrypoint script gives exit 0
    to end the main script process.
3. trap command is waiting for EXIT.  On EXIT, if 'launchfile' is
    present, will delete launchfile and restart the app.

Using this approach, the main app icon will always remain during
the nwjs processes.

Change-Id: I38685b184ecd597a9b80c40e6e6e2837f51f2e1c