Fix XDTestSuite.sh to use ssh
Move to caddy forces the use of https because it does not support
http in product installs.  The http-based test for XD needs to connect
to the instance using https, which forces the test server to recieve
updates using https.  This means the commands to start and monitor
the tests must use https.

Change-Id: I9056a73850e418df0446d16c559a457d2eb4a89d