xce: add forceinstall option to main executable
Right now XPE script will only open installer if there is no
xdpce container on host, or if the latest xdpce container is
for a build not included in the app.  this adds a forceinstall
option, such that if it is passed when the script is invoked,
will always run installer.

This is needed for the 're-install' option being added to
nwjs menu.

Change-Id: I92c9bc776b2342205f1111b4ebf120c77ca5e096