Use devTemplate.json instead of azuredeploy.json
devTemplate.json is a copy of mainTemplate.json with 1 modification:
creating 1 IP per VM.

Previosuly azuredeploy.json was relying on node0 to export the NFS
server, however, the azure-dns wasn't reliable enough for every node to
resolve node0's hostname in order to mount the NFS.

The solution to this was to mount a CIFS that's globally visible. This
was already done in mainTemplate.json.

We wanted to use mainTemplate.json directly, but because it creates only
1 public IP per cluster, and the conditional to create 1 public IP per
VM for PoseidonTestSuite was getting really cumbersome, so we forked off
devTemplate.json instead.

Change-Id: I9401869ebe8dce5889047dcd5c93c59b7382910a