xpe: Host home mount to match xcalar changes
The default data target for builds having "Xcalar Design" product
family in the license file (X-CE and X-EE), are going to be set
in the Xcalar backend code

Review: https://gerrit.int.xcalar.com/#/c/9575/

As a result, remove section of setup where we were creating a
data target.  Also, name of mount point needs to match what is
the default in XCE, so adding a comment so that this string does
not get changed in future revisions.

Change-Id: Iba8d50783bafddf88d8bf7e922c6fde8abd25039