GR: Detect corruption edge case
Due to the SSE optimizations the compiler assumes 16 byte alignment for
at least some allocations.  To fulfill this assumption requires opening
up a region of zero to 15 bytes before the guard page.  This typically
happens with odd size allocations and/or alignments.  So given a
sufficiently small overrun in such cases, we will not detect it with the
MMU alone.  This code will at least detect the corruption after the
fact.

Change-Id: I5b15260869555a55b831ca5bf8341e627b1b83b1