installerPath ignored for local perf test
As it is, we're running perf test locally, in which case installerPath
is ignored anyway.

In future, we can augment this job to run the perf test on a remote
node. In which case, we need to also specify the --remote flag, so
additional intelligence will need to be wired into this Jenkins job.

Change-Id: Ia598d5d1ca5d636770ad299065911a1347e815a1