xpe: more actions possible when app opens
Summary: Previously app opens only installer or XD as an nwjs
process.
with this change now app can open either installer, uninstaller,
revert tool, or XD as an nwjs process

Determines to open tool by following: (ordered by precedence):
    1. cmdline param to script (install, uninstall, revert, xd)
    2. launchfile content (install, uninstall, revert, xd)
    3. if none:
        (a) if xdpce:latest not present or not same as build packaged
            with this app --> installer
        (b) else --> xd

About launchfile mentioned in 2.:
The file is set by server api which is called when the user closes
an nwjs window under certain circumstances (for ex: if they click re-install,
revert, or uninstall from custom menu while XD running, or when selecting
'launch' button in installer at end of install process);
the file has contents only specifying which next tool to run
(install, uninstall, revert, xd)
This launchfile was being used before, but previously the script was only
testing for existence of file; this is because before it was only set in one
scenario, as an empty file at end of install process if user pressed 'launch' button;
Script would just restart app and launch XD on new start if the file was detected.

Change-Id: I4a0306e9bed65a9f02322d167fd3e4b5dc5c5b4d