xpe: start Docker on app strt unless NODOCKERSTART
Each time the app starts, it will check if Docker daemon available.
If not, will launch a GUI that attempts to start Docker.
(The GUI code and API endpoints it calls are in xcalar-gui project,
but the endpoints call the new functions in local_mac_installer.sh
introduced in this change: cmd_ensure_docker_up,
cmd_start_wait_docker, and cmd_check_docker)

If env variable NODOCKERSTART is set, Docker will not be started.
(In this case the GUI is erroring with this reason)

Change-Id: Ibfcebcf5377ad308e909873a2f006c261bd21e31