xpe: Allow option to build Grafana or not
Currently every build is also creating a Grafana docker container,
and including the saved image in the installer tarball.
On the host side, a new Grafana container is always being installed.
However, in official product do not want Grafana included (it should
be being included only for internal or dev distributions)

This change allows for this to be an option at build time.

Here is how this change works, and the order the files will be
utilized:

(Context:: the Jenkins job 'XcalarPersonalEditionBuilder'
now has a boolean option BUILD_GRAFANA)

1. xpe/jenkins/XcalarPersonalEditionRunner.sh -->
   This shell script gets called directly by Jenkins job XcalarPersonalEditionBuilder.
   If BUILD_GRAFANA env variable present, only then will it create
   grafana container.

2. xpe/scripts/makeapp.sh -->
   This script is called by 1. as part of build, after containers have
   been made.  It generates the actual Mac app.
   If BUILD_GRAFANA env variable is present, it will touch a blank
   file in the Contents/MacOS dir of the app (where the main app
   executable lives).  This will be a way for the app to detect
   at install time, that Grafana is intended to be installed

3. xpe/scripts/Xcalar Design -->
   This is the main app executable; it gets run every time the app is
   double clicked.
   It will check if the file Contents/MacOS/.grafana exists in the app
   (the file touched in 2.)
   If so, it will export an env variable, INSTALL_GRAFANA.
   This env variable can now be detected by the js running the server
   on the host, and the helper scripts run during install time

4. xpe/scripts/local_installer_mac.sh -->
   This is a helper script which includes functions which are called
   by the server that gets run on the host.  (The server code is in xcalar-gui)
   Some of the functions now check if INSTALL_GRAFANA env present
   (set in 3.), and if so take diff action than before
   (for example, in the function which is called by the server during
   install to load the docker images, if INSTALL_GRAFANA is set,
   will make sure the grafana tarball is present, else this whole
   function will return non-0)

Note: The javascript in the installer GUI will also be checking for
INSTALL_GRAFANA env variable (set in 3.).  Only if it is present
will the function cmd_create_grafana in local_installer_mac.sh  get
called; hence why there is no check in this function for INSTALL_GRAFANA

Change-Id: I17f9085af73bf4e799a8a6f8fac25e9413ffc778