xpe: allow dev/nondev, separate from grafana op
Change here allowed to make Grafana inclusion a build time option:
https://gerrit.int.xcalar.com/#/c/9953/

Also will like to make dev/non-dev builds which will have an
influence on certain aspects of the javascript and GUI.
Right now, the guis just looking for if Grafana included, and if
so, are assuming its a dev build, and making the dev changes.
However, this change allows for dev/non-dev build to be a completely
separate option.  This way, you could have a build without the
dev javascript, but which still can include Grafana.

Change-Id: I2c32899a5eabbf8768e1b1a898f4b129c29e4810