xpe: seperatate Docker start and Docker wait functions
Current method for bringing up Docker, starts
Docker app and then waits for it to come up.

a new Api will need to wait for Docker daemon to be
available, but will not want to call Docker start.
Therefore, break the bringing up Docker in to two
functions: one for starting Docker, another for
waiting.  Now the new API can call this single
wait function, and the existing APi will continue to
work.

Change-Id: Ie3b067e6839e52209f7a22ea378687bc01dd3af2