xpe: cleanup Docker images even on build failure
The Xcalar docker images and containers created during this Jenkins
job need to be removed, or on future runs done by the same Jenkins slave,
 docker save will include those leftover images.

Currently the images are being removed, but only if the script succeeds and gets
to the end.  Because of this, I've been having to ssh in to the jenkins
slave after failures to clean things up.

This adds a trap command which will execute on every exit, regardless
of success or failure.  TODO: Handle abort via Jenkins - haven't been
 able to catch this

Change-Id: Id6a6f220b1eb5f51d1e8f1e7b089311637ac0efd