xpe: use named params for GUI force (bug fix)
The app entrypoint determines which GUI should be opened by nwjs.
Currently, cmd param can override the default behavior to force a
particular GUI.  Examples:

  bash Xcalar\ Design install   // should open installer
  bash Xcalar\ Design uninstall // should open uninstaller
  bash Xcalar\ Design xd        // should open xd
  bash Xcalar\ Design revert    // should open image management tool

However, the script is only just checking that there IS a cmd
 param, and not validating anything about the param.
I have discovered that for older versions of mac, depending on how the
 app is opened (finder, open cmd, etc.) the operating system
 will run the app entrypoint and specify cmd params to it.
When the OS runs with one of these params, the entrypoint thinks
 it's trying to force a particular GUI, but since the param is not anything
 it recognizes, this causes a blank nwjs window to open.

Therefore, am switching to use named param.  -a will allow to force a GUI,
 all other cmd params are not being looked at.  It'll now work like this:

  bash Xcalar\ Design -a install   // opens installer
  bash Xcalar\ Design -a uninstall // opens uninstaller
  bash Xcalar\ Design -a xd        // opens xd
  bash Xcalar\ Design -a revert    // opens image management tool

Change-Id: Id2e3d65dbb58f3be75079e81c99c78f34b322935