ovirttool: modularize VM power on
When running with --delete option, will try to power on the VM to
 release its IP.  Tries to power on the VM by just checking if the VM
 is not up and if not, powers it on.
If you try to power on a VM in a 'powering up' state, it will fail.
So if you try to run with --delete on a VM that's just been powered on,
 and is in a 'powering up' state, then currently the script will fail
 (as 'powering up' is not 'up')
Instead, check if a VM is in a post-powering up state, and only  if its
 not should you attempt to start the service.
In fixing this, ended up modularizing things a big.

Change-Id: I1b89834abdc995994d0bad9ad14101856df75cfb