fix exitCode handling for XD tests
Previous syntax set exitCode=1 and then used an if/elif/else to
perform a series of commands of the form:
npm test -- sqlTest https://localhost:8443 || exitCode=$?
and ended with:
exit $exitCode

The intention of this construct appeared to be:
1) The test fails by default (exitCode=1)
2) Use if/elif/else to choose the necessary testing depending on how
the script is called
3) Run the test, and capture the return code of the test.

The logic in 1) and 3) is subtlely wrong.
exitCode=1
npm test -- sqlTest https://localhost:8443 || exitCode=$?
The nature of the || is such that exitCode is only updated if
the npm command returns a non-zero return code because of conditional
execution (a || b and a is true ($?=0) then b never runs).

This change corrects that problem by removing the <test command> || exitCode=$?
syntax and collecting $? after the correct test is run.  This lets
the exitCode be initialized to 1 and still always collect the correct
return value.

It also removes some kill statements that accomplished nothing.

Change-Id: I9e878aec50881fc3087fd436ff0aa3de50ff11d0