XD-1721, XD-1722, XD-1723, TOOLS-187 - fix the last XDTestSuite to use the new location of xcalar-sqldf.jar (2nd try)
The change https://gerrit.int.xcalar.com/#/c/15216 altered the location and
method by which xcalar-sqldf.jar should be bundled for testing and distribution.

This change alters XDTestSuite.sh to use the correct method for obtaining
xcalar-sqldf.jar according to that change.

An earlier version of this change redirected the output of download-sqldf.sh to
the file location.  This is breaks download-sqldf.sh.  The destination needs to
be the first argument.

This version also detects whether download-sqldf.sh can be used if an old branch does
not have the change ported.  If not, it employs the earlier method of getting xcalar-sqldf.jar.

Change-Id: Ibb9a087bba73639f1ba2fb780191b30ea06222d0