Fix ovirttool.py dependency on freenas2
ovirttool has two major types of dependencies on using freenas2 with
the VMs it produces.  It explicitly refers to freenas2 by name or IP
address in ovirttool.py.  The setup logic in ovirttool.py also implicitly
requires that /netstore be already set up to use freenas2 on any VM
templates that it uses to create virtual machines.

The latter requirement comes in two forms: incorrect logic for establishing
the correct /netstore mountpoint should an existing path on cluster
shared storage not exist, and puppet agent failure while reconfiguring
the /netstore mount (to freenas3) because Xcalar is using it.

This change converts all references to freenas2 to freenas3 in
ovirttool.py.  It shuts down Xcalar before running the puppet agent,
so the two services do not interfere with each other.  Finally, it
changes the shared cluster storage mounting logic during cluster
configuration to correctly manually mount the freenas3 /netstore export
if it is not already present.

Change-Id: I616619fc17e2b77963b92be0a3f9ad8596515eb2