SQL-754 sqlscaltest xccli -c version timeout
The root cause is that ssh to cluster and running xccli -c version
as default user Jenkins got permission deny. So the command cannot
get the successful response. The fix is specifying
the user as Xcalar when ssh to the cluster and run
the command, which could solve the permission deny problem.

Change-Id: I92646c1fde6a25eeada8074f7c4c815f69ae9862