TOOLS-600 Azure: Remove systemd units from cloud deployment
Previously the azure setup code had its own systemd
unit files. This goes back to a time before we used
systemd units on EL7 in the product, but had to use
it in cloud to get proper ordering dependencies of
the units to play nice.

Also simplify the cluster code to not re-specify all
the params that should be in the provided file. You
can still specify them, but the defaults in the
script don't clobber the defaults in the json anymore.

Add support for ephemeral OS disks (not on by default)
These are free disks you can have access to and provide
very fast i/o. Great for test setups.

Change-Id: Ie3d45e3c22356800ed351f02295510b3ac4204e4