Fix unmount of /netstore command
A previous fix to ovirt in https://gerrit.int.xcalar.com/#/c/16120/ caused
/netstore be unmounted before correctly mounting that location from freenas3
using this command:

mountpoint -q /netstore && umount /netstore

If /netstore is mounted, this works fine.  If it is not, the mountpoint part
of the command returns a non-zero status, causing run_ssh_cmd to fail.

This change alters the command to return status 0 if /netstore is not mounted
or /netstore is mounted and the unmount is successful.

Change-Id: Ib2a9e30bb45664b2f1663618e04b22eeeae02257