ENG-566, ENG-568, ENG-569 - Set up Passport auth framework using aws-serverless-express
This provides an initial implementation of an AWS Lambda that uses Passport
and aws-serverless-express to create an expServer-like login process for AWS.
Details on how to build and use the Lambda can be found here:

https://xcalar.atlassian.net/wiki/spaces/EN/pages/68943873/Building+the+Xcalar+SaaS+Authentication+Lambda

The current patch set provides implementation for the following endpoints:

- login
- logout

There are a number of things that could be done better.  Some code left over
from the aws-serverless-express code needs to stripped away.  It needs to be
better integrated with the Xcalar SaaS overall build/deployment.  Some
DynamoDB table security issues need to be resolved.  A status endpoint needs
to be implemented, and corresponding python status code needs to be implemented.

Also adds a properly configured role that should allow EC2 instances to access the
Session table.

It now passes through all the credentials required for a python lambda to perform
a get_credentials_for_identity boto3 python call.

It also allows CORS requests.

Change-Id: Ic2689cda88727fabd841c81c46b9a9dc214d37e8