ENG-4817 - Allow SaaS Auth Lambda to issue cookies for other domains
Xcalar SaaS is now presumed to need to run across multiple locations.
To do that, multiple session cookies will be needed.

To do this, this change adds two new lambda endpoints.  The /creds
endpoint will copy Xcalar session cookies for two groups of cookies
domains.   The first is supplied in body.domains -- this handles
dynamic domains (such as user clusers) that the infrastructure knows
about at runtime.   The second is supplied in an environment variable
in the Lambda itself -- these are infrastruction domains that should
be added to all request automatically.

The POST body in the first case just needs to include a JSON list
called domains:

{ "domains": ["foo.org", "bar.com"] }

In the second case, the environment varable separates domains with
colons:

"smith.net:jones.edu"

The "/status" endpoint provides session information much that already
provided by the expServer /sessionStatus path:

{"user":false,"admin":true,"loggedIn":true,"emailAddress":"thaining@xcalar.com",
"firstName":"Ted","username":"thaining@xcalar.com","timeout":1800}

The "/status" path is unsecured and will always return data.  The loggedIn
member indicates whether or not the user is logged in.

Change-Id: I99dde583e248c486c8e0c031a1831c7839a9f6b0