ENG-4817, ENG-4859 - CORS fixes to Lambda and removal of third party token code
Changes in https://gerrit.int.xcalar.com/#/c/17546 did not work during testing.
So, this change removes the /creds endpoint from the Lambda.

This change also fixes up CORS problems for the Lambda stack -- it properly sets
the CORS configuration in the swagger file, and then it also properly configures
the CORS response in the nodejs code.   The two have to agree because the AWS
infrastructure provides the OPTIONS response that are used to verify the origin,
but the nodejs stack actually crafts the response headers that are returned at
the end of the request.

Change-Id: I18a4957f48a42af345175b43c481afd94b69da0b