ENG-4862 - Move the build of the Auth Lambda to Jenkins
This change automates the deployment of the Auth Lambda with a Jenkins
job called DeployAuthLambda.  The process creates a bucket to hold the
lambda, configures the lambda source with a series of inputs, does some
npm package management, packages the lambda, and deploys it.

A description of the code configuration inputs can be found on Confluence:

https://xcalar.atlassian.net/wiki/spaces/EN/pages/68943873/Building+the+Xcalar+SaaS+Authentication+Lambda

Along with a script to deploy the Lambda, another script is capable of
removing it.  It takes the bucket name and the stack name from the
Lambda deployment and removes both of these resources.

Change-Id: I4c97cd9c0972403e1a8fc2f2ecdf027ec4f94390