ENGR 5001 - Auth lambda not work with cookie cache
This bug concerns the following SaaS auth exchange:

- login with the auth lambda (cookie set A)
- passing the session info from the auth lambda to
  the cluster expServer (cookie set B)
- logging out via the expServer (cookie set B)
- checking the login status with the auth
  lambda (cookie set A)

Checking the login status with the lambda caused an
error.  The root cause of the error was in complete
DynamoDB session table rows that were saved/left in
the session table.

This change extends fixes to the expServer that cause
the malformed row in the session table by doing some
better error checking around the auth lambda logout
and sending cookie expiration requests back to the calling
client for both the /logout and /status calls.

Change-Id: Iedc33f9df161c1fa9e3d61a26587973d5c09d961