ENG-501 - Improve session table security by checking cluster owner and making secrets private to auth lambda
In order to improve access security, requests to Xcalar SaaS clusters
should only allow access to requests from sessions that belong to the
cluster owner.

This requires the expServer to obtain the cluster owner at startup.
In order to do that, the ec2 IAM role needs to allow DescribeTags.

This change adds DescribeTags to the policy.

Change-Id: I8bdc9a2dd49f619c5218e9d64d0e5f9d0a40bba0