ENG-5032 - Automate configuration of the expServer for SaaS during AMI build
The previous attempt to config the expServer in
https://gerrit.int.xcalar.com/#/c/17815 failed in at least two
ways. It was trying to configure the expServer at build time
rather than bootstrap. It was trying to use a config lambda when
it should be using SSM parameters.

This change removes the configuration step from the AMI build.  It
also removes references to the lambda and its ARN.

To handle the passing of lambda configuration information, the
DeployAuthLambda job now stores suitable configuration information
at the path /xcalar/cloud/auth/{AuthStackName}.

This means that the DeployAuthLambda CLOUDFORMATION_STACK_NAME must
now be provided to the DeployShells and UpdateShells Jenkins Jobs.
The shell scripts that run these jobs now expect this value in the
parameter AUTH_STACK_NAME.

Changes to the xdp-standard.template.j2 now make certain steps
for cloud auth dependent on the enableCloudAuth variable in extra.yaml.
The AUTH_STACK_NAME will also be placed in the file
/var/lib/cloud/instance/cloudauth.env on the ec2 nodes.

The presence of this file will cause the bootstrap script (user-data.sh)
to read the file, get the parameters and idempotently put the config
in /etc/default/xcalar.

Change-Id: I93d670b3e72a98bce62b06b193a9c138c29aa42f