ENG-501 - Improve session table security by moving AWS credentials to separate table
The Session table used by Xcalar contains entries created by the
auth lambda, and read by the expServer on all EC2 nodes.  This
makes the table a security liability because programs on the EC2
nodes can potentially see all user sessions.   A few things make
seeing all the sessions difficult:

- session ids are randomly assigned
- the session id namespace is huge
- bandwidth between the EC2 nodes and dynamo db is limited

This changes that scheme by moving the cognito creds to a separate
table, organized by username and sessionID.

The check_user_credentials routine for the python lambdas is also
updated to perform the necessary retrieval from the new table.

The creds table uses a composite primary key created from the
Cognito account username and the session id.  If we wish to
eventually use row-level access on the creds table, we can
switch the username to a property like Cognito userId.

Change-Id: I05f30a852a3c3d8db06d0b0915e2570a41a8f6ad