ENG-5396 - Cannot connect to my workbook in Jupyter in the Developer Cloud product
This problem was traced to authentication cookie domain issues.
To make cookies work for all lambdas, they were set to a domain
.execute-api.${region}.amazonaws.com when they are created.  The
problem is that the domain is a saved to the dynamodb session table
and we want to the cookies to be used in two cookie domains:
the lambdas and the EC2 clusters.  Since the table is shared between
them, the expServer on EC2 was creating cookies with the lambda domain.
This caused the cookies to be ignored by the SDK cookie cache and
SDK auth fails.

This change alters the mechanism for supplying lambda cookies.
Instead of setting a saved default, it creates additional lambda
domain cookies that should work for all lambdas.

It also makes some minor consistency fixes to lambda error messaging.

Change-Id: Ibe61a94ecdf18cec47c8c597619c3b193f1f9fe6