ENG-500 - fix /status cookie handling in the auth lambda
Changes in https://gerrit.int.xcalar.com/#/c/18226 altered the
handling of cookies by the auth lambda.  Code added there caused
the /status route to attempt to return valid session cookies if
the session was logged in or not.  This was incorrect, and caused
a status 500 failure because data required to create cookies did
not exist.

This change fixes the error, causing valid cookies to be returned
if the user is logged in and expired cookies if not.

Change-Id: Ib562d9939ef9c2a8b2e54ab8821b46f275bf4bb0