ENG-5294 Don't swap to HDD
Running SystemTests against a system with more swap than what we
recommend/configure for clients isn't a good test of system stability.

Additionally, if we exhaust memory and SSD swap we shouldn't start
swapping to an HDD. This causes the system to move extremely slowly and
may mask/prolong the time it takes for us to realize a memory leak
exists.

Tests should be run on a system similar to an actual deployment.

Change-Id: I5fcd859b1991dac44c0aa9d381df7fd60b1fbfac