ENG-6123 - XDUnitTest, XCEFuncTest, XDEndToEndTest and XDFuncTest fails due to improper default.cfg file
XDTestSuite.sh depends on XCE config determined by $XLRDIR/src/bin/util/xcalar/compute/util/cluster.py.
Depending on configuration, this can resolve to four different locations:
1) the value of XCE_CONFIG
2) /etc/xcalar/default.cfg
3) $XLRDIR/src/data/{hostname}.cfg
4) $XLRDIR/src/data/test.cfg

In order to better insure which location is chosen, this change sets XCE_CONFIG to
a value, so that XCE_CONFIG is chosen if it exists.

Change-Id: I5dc2fbdd5fe6d3df4edbb9593dd76757f5a0c64b