ENG-6555 schema checksum and parity
Computing JSON and CSV schema checksums on the Lambda
side. We still send the entire schema to client, but
if client would like to do fast comparisons the
checksums can be used. In case of JSON, column order
is not strict and BigQuery allows if there is
inconsistent column order as it's still the same
schema. So we use sort, normalize the schema and compute
checksum in case of JSON but do not sort in case
of CSV as column order is strict. For trailing columns
scenario in CSV special logic needs to be used and
checksum cannot be used.

Change-Id: I247cca58ae0899ddaa228b36639827edfdb8a4e7