ENG-7276 - PoseidonSystemTest fails with "connection refused" testing 2.0.5-RC22
Changes in https://gerrit.int.xcalar.com/c/xcalar/xcalar-infra/+/20129
corrected issues with commands to start and stop Xcalar for testing.

The command used to determine if Xcalar was running was left incomplete
because it only determined if the unit running Xcalar was started, not
whether a usrnode process was actually running.

This change fixes that by adding commands and conditions that detect
the running usrnode.  The newer systemd unit hierarchy can do this
cleanly.  Where the original xcalar.service unit is used, I must
employ a hack and call the init.d script.

Change-Id: I972cfa87cfcade58f8052e7e52a5dd9fb175b7dc