ENG-7595 event monitor lambda function
We want a small service to record custom events such as installation
begining/end. It was decided to implement it as a lambda function
writing to a dynamodb.

This is a PoC, supporting an /events Api endpoint that allows GETs and
POSTs, former to be used for listing events and latter to submit new
ones.

Change-Id: Iaafd8fefdd106e2f1507a054e4de049c96d292ba