ENG-7762 setup ovirt prom scraper
We are missing data from the ovirt servers (we have the VMs). There's an
ovirt prom scraper so we're setting that up rather than trying to add
node_exporter to the servers.

the ovirt scraper was not working out of the box, there's a problem with
disk stats, so I made a couple changes to work around it and pushed to
our registry a modified image.

Change-Id: Icd320f3b74a24f12a30ef649f115e546aefea4f4